# packages/sdk/openapi.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "openpatent",
    "description": "openpatent api",
    "version": "1.0.0"
  },
  "paths": {
...
    "/session/{sessionID}/message/{messageID}": {
      "get": {
        "operationId": "session.message",
        "parameters": [
          {
            "in": "query",
            "name": "directory",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "sessionID",
            "schema": {
              "type": "string"
            },
            "required": true,
            "description": "Session ID"
          },
          {
            "in": "path",
            "name": "messageID",
            "schema": {
              "type": "string"
            },
            "required": true,
            "description": "Message ID"
          }
        ],
        "summary": "Get message",
        "description": "Retrieve a specific message from a session by its message ID.",
        "responses": {
          "200": {
            "description": "Message",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "info": {
                      "$ref": "#/components/schemas/Message"
                    },
                    "parts": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Part"
                      }
                    }
                  },
                  "required": [
                    "info",
                    "parts"
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BadRequestError"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotFoundError"
                }
              }
            }
          }
        },
        "x-codeSamples": [
          {
            "lang": "js",
            "source": "import { createopenpatentClient } from \"@openpatent-ai/sdk\n\nconst client = createopenpatentClient()\nawait client.session.message({\n  ...\n})"
          }
        ]
      }
    },
...
    "/config/providers": {
      "get": {
        "operationId": "config.providers",
        "parameters": [
          {
            "in": "query",
            "name": "directory",
            "schema": {
              "type": "string"
            }
          }
        ],
        "summary": "List config providers",
        "description": "Get a list of all configured AI providers and their default models.",
        "responses": {
          "200": {
            "description": "List of providers",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "providers": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Provider"
                      }
                    },
                    "default": {
                      "type": "object",
                      "propertyNames": {
                        "type": "string"
                      },
                      "additionalProperties": {
                        "type": "string"
                      }
                    }
                  },
                  "required": [
                    "providers",
                    "default"
                  ]
                }
              }
            }
          }
        },
        "x-codeSamples": [
          {
            "lang": "js",
            "source": "import { createopenpatentClient } from \"@openpatent-ai/sdk\n\nconst client = createopenpatentClient()\nawait client.config.providers({\n  ...\n})"
          }
        ]
      }
    },
    "/provider": {
      "get": {
        "operationId": "provider.list",
        "parameters": [
          {
            "in": "query",
            "name": "directory",
            "schema": {
              "type": "string"
            }
          }
        ],
        "summary": "List providers",
        "description": "Get a list of all available AI providers, including both available and connected ones.",
        "responses": {
          "200": {
            "description": "List of providers",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "all": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "api": {
                            "type": "string"
                          },
                          "name": {
                            "type": "string"
                          },
                          "env": {
                            "type": "array",
                            "items": {
                              "type": "string"
                            }
                          },
                          "id": {
                            "type": "string"
                          },
                          "npm": {
                            "type": "string"
                          },
                          "models": {
                            "type": "object",
                            "propertyNames": {
                              "type": "string"
                            },
                            "additionalProperties": {
                              "type": "object",
                              "properties": {
                                "id": {
                                  "type": "string"
                                },
                                "name": {
                                  "type": "string"
                                },
                                "family": {
                                  "type": "string"
                                },
                                "release_date": {
                                  "type": "string"
                                },
                                "attachment": {
                                  "type": "boolean"
                                },
                                "reasoning": {
                                  "type": "boolean"
                                },
                                "temperature": {
                                  "type": "boolean"
                                },
                                "tool_call": {
                                  "type": "boolean"
                                },
                                "interleaved": {
                                  "anyOf": [
                                    {
                                      "type": "boolean",
                                      "const": true
                                    },
                                    {
                                      "type": "object",
                                      "properties": {
                                        "field": {
                                          "type": "string",
                                          "enum": [
                                            "reasoning_content",
                                            "reasoning_details"
                                          ]
                                        }
                                      },
                                      "required": [
                                        "field"
                                      ],
                                      "additionalProperties": false
                                    }
                                  ]
                                },
                                "cost": {
                                  "type": "object",
                                  "properties": {
                                    "input": {
                                      "type": "number"
                                    },
                                    "output": {
                                      "type": "number"
                                    },
                                    "cache_read": {
                                      "type": "number"
                                    },
                                    "cache_write": {
                                      "type": "number"
                                    },
                                    "context_over_200k": {
                                      "type": "object",
                                      "properties": {
                                        "input": {
                                          "type": "number"
                                        },
                                        "output": {
                                          "type": "number"
                                        },
                                        "cache_read": {
                                          "type": "number"
                                        },
                                        "cache_write": {
                                          "type": "number"
                                        }
                                      },
                                      "required": [
                                        "input",
                                        "output"
                                      ]
                                    }
                                  },
                                  "required": [
                                    "input",
                                    "output"
                                  ]
                                },
                                "limit": {
                                  "type": "object",
                                  "properties": {
                                    "context": {
                                      "type": "number"
                                    },
                                    "output": {
                                      "type": "number"
                                    }
                                  },
                                  "required": [
                                    "context",
                                    "output"
                                  ]
                                },
                                "modalities": {
                                  "type": "object",
                                  "properties": {
                                    "input": {
                                      "type": "array",
                                      "items": {
                                        "type": "string",
                                        "enum": [
                                          "text",
                                          "audio",
                                          "image",
                                          "video",
                                          "pdf"
                                        ]
                                      }
                                    },
                                    "output": {
                                      "type": "array",
                                      "items": {
                                        "type": "string",
                                        "enum": [
                                          "text",
                                          "audio",
                                          "image",
                                          "video",
                                          "pdf"
                                        ]
                                      }
                                    }
                                  },
                                  "required": [
                                    "input",
                                    "output"
                                  ]
                                },
                                "experimental": {
                                  "type": "boolean"
                                },
                                "status": {
                                  "type": "string",
                                  "enum": [
                                    "alpha",
                                    "beta",
                                    "deprecated"
                                  ]
                                },
                                "options": {
                                  "type": "object",
                                  "propertyNames": {
                                    "type": "string"
                                  },
                                  "additionalProperties": {}
                                },
                                "headers": {
                                  "type": "object",
                                  "propertyNames": {
                                    "type": "string"
                                  },
                                  "additionalProperties": {
                                    "type": "string"
                                  }
                                },
                                "provider": {
                                  "type": "object",
                                  "properties": {
                                    "npm": {
                                      "type": "string"
                                    }
                                  },
                                  "required": [
                                    "npm"
                                  ]
                                }
                              },
                              "required": [
                                "id",
                                "name",
                                "release_date",
                                "attachment",
                                "reasoning",
                                "temperature",
                                "tool_call",
                                "limit",
                                "options"
                              ]
                            }
                          }
                        },
                        "required": [
                          "name",
                          "env",
                          "id",
                          "models"
                        ]
                      }
                    },
                    "default": {
                      "type": "object",
                      "propertyNames": {
                        "type": "string"
                      },
                      "additionalProperties": {
                        "type": "string"
                      }
                    },
                    "connected": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      }
                    }
                  },
                  "required": [
                    "all",
                    "default",
                    "connected"
                  ]
                }
              }
            }
          }
        },
        "x-codeSamples": [
          {
            "lang": "js",
            "source": "import { createopenpatentClient } from \"@openpatent-ai/sdk\n\nconst client = createopenpatentClient()\nawait client.provider.list({\n  ...\n})"
          }
        ]
      }
    },
...
    "/tui/control/response": {
      "post": {
        "operationId": "tui.control.response",
        "parameters": [
          {
            "in": "query",
            "name": "directory",
            "schema": {
              "type": "string"
            }
          }
        ],
        "summary": "Submit TUI response",
        "description": "Submit a response to the TUI request queue to complete a pending request.",
        "responses": {
          "200": {
            "description": "Response submitted successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "boolean"
                }
              }
            }
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {}
            }
          }
        },
        "x-codeSamples": [
          {
            "lang": "js",
            "source": "import { createopenpatentClient } from \"@openpatent-ai/sdk\n\nconst client = createopenpatentClient()\nawait client.tui.control.response({\n  ...\n})"
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "UserMessage": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "sessionID": {
            "type": "string"
          },
          "role": {
            "type": "string",
            "const": "user"
          },
          "time": {
            "type": "object",
            "properties": {
              "created": {
                "type": "number"
              }
            },
            "required": [
              "created"
            ]
          },
          "summary": {
            "type": "object",
            "properties": {
              "title": {
                "type": "string"
              },
              "body": {
                "type": "string"
              },
              "diffs": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/FileDiff"
                }
              }
            },
            "required": [
              "diffs"
            ]
          },
          "agent": {
            "type": "string"
          },
          "model": {
            "type": "object",
            "properties": {
              "providerID": {
                "type": "string"
              },
              "modelID": {
                "type": "string"
              }
            },
            "required": [
              "providerID",
              "modelID"
            ]
          },
          "system": {
            "type": "string"
          },
          "tools": {
            "type": "object",
            "propertyNames": {
              "type": "string"
            },
            "additionalProperties": {
              "type": "boolean"
            }
          },
          "variant": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "sessionID",
          "role",
          "time",
          "agent",
          "model"
        ]
      },
...
      "Message": {
        "anyOf": [
          {
            "$ref": "#/components/schemas/UserMessage"
          },
          {
            "$ref": "#/components/schemas/AssistantMessage"
          }
        ]
      },
...
      "Part": {
        "anyOf": [
          {
            "$ref": "#/components/schemas/TextPart"
          },
          {
            "type": "object",
            "properties": {
              "id": {
                "type": "string"
              },
              "sessionID": {
                "type": "string"
              },
              "messageID": {
                "type": "string"
              },
              "type": {
                "type": "string",
                "const": "subtask"
              },
              "prompt": {
                "type": "string"
              },
              "description": {
                "type": "string"
              },
              "agent": {
                "type": "string"
              },
              "command": {
                "type": "string"
              }
            },
            "required": [
              "id",
              "sessionID",
              "messageID",
              "type",
              "prompt",
              "description",
              "agent"
            ]
          },
          {
            "$ref": "#/components/schemas/ReasoningPart"
          },
          {
            "$ref": "#/components/schemas/FilePart"
          },
          {
            "$ref": "#/components/schemas/ToolPart"
          },
          {
            "$ref": "#/components/schemas/StepStartPart"
          },
          {
            "$ref": "#/components/schemas/StepFinishPart"
          },
          {
            "$ref": "#/components/schemas/SnapshotPart"
          },
          {
            "$ref": "#/components/schemas/PatchPart"
          },
          {
            "$ref": "#/components/schemas/AgentPart"
          },
          {
            "$ref": "#/components/schemas/RetryPart"
          },
          {
            "$ref": "#/components/schemas/CompactionPart"
          }
        ]
      },
...
      "BadRequestError": {
        "type": "object",
        "properties": {
          "data": {},
          "errors": {
            "type": "array",
            "items": {
              "type": "object",
              "propertyNames": {
                "type": "string"
              },
              "additionalProperties": {}
            }
          },
          "success": {
            "type": "boolean",
            "const": false
          }
        },
        "required": [
          "data",
          "errors",
          "success"
        ]
      },
      "NotFoundError": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "const": "NotFoundError"
          },
          "data": {
            "type": "object",
            "properties": {
              "message": {
                "type": "string"
              }
            },
            "required": [
              "message"
            ]
          }
        },
        "required": [
          "name",
          "data"
        ]
      },
...
      "Provider": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "source": {
            "type": "string",
            "enum": [
              "env",
              "config",
              "custom",
              "api"
            ]
          },
          "env": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "key": {
            "type": "string"
          },
          "options": {
            "type": "object",
            "propertyNames": {
              "type": "string"
            },
            "additionalProperties": {}
          },
          "models": {
            "type": "object",
            "propertyNames": {
              "type": "string"
            },
            "additionalProperties": {
              "$ref": "#/components/schemas/Model"
            }
          }
        },
        "required": [
          "id",
          "name",
          "source",
          "env",
          "options",
          "models"
        ]
      },
...
    }
  }
}